requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export --format requirements-txt --no-hashes
-e .
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.2
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
audioop-lts==0.2.1 ; python_full_version >= '3.13'
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6 ; sys_platform == 'win32' or platform_system == 'Windows'
exceptiongroup==1.3.0
fastapi==0.115.12
fastmcp==2.5.1
ffmpy==0.5.0
filelock==3.18.0
frozenlist==1.6.0
fsspec==2025.5.1
gradio==5.31.0
gradio-client==1.10.1
groovy==0.1.2
h11==0.16.0
hf-xet==1.1.2 ; platform_machine == 'aarch64' or platform_machine == 'amd64' or platform_machine == 'arm64' or platform_machine == 'x86_64'
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.32.2
idna==3.10
jinja2==3.1.6
markdown-it-py==3.0.0
markupsafe==3.0.2
mcp==1.9.1
mdurl==0.1.2
multidict==6.4.4
numpy==2.2.6
openapi-pydantic==0.5.1
orjson==3.10.18
packaging==25.0
pandas==2.2.3
pillow==11.2.1
propcache==0.3.1
pydantic==2.11.5
pydantic-core==2.33.2
pydantic-settings==2.9.1
pydub==0.25.1
pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
pyyaml==6.0.2
requests==2.32.3
rich==14.0.0
ruff==0.11.11 ; sys_platform != 'emscripten'
safehttpx==0.1.6
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sse-starlette==2.3.5
starlette==0.46.2
tomlkit==0.13.2
tqdm==4.67.1
typer==0.16.0
typing-extensions==4.13.2
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
uvloop==0.21.0 ; platform_python_implementation != 'PyPy' and sys_platform != 'cygwin' and sys_platform != 'win32'
watchfiles==1.0.5
websockets==15.0.1
yarl==1.20.0
